Convert retrieved values to double
We rely on MatlabControl to extract data from MATLAB and pass it to our
current JVM (if different). Unfortunately, the MatlabNumericArray class
can only wrap MATLAB double arrays. When unwrapped arrays are returned
directly their dimensionality is lost (a 1D array is returned).

So we must convert to double arrays using the MATLAB "double" function
to ensure data is wrappable by MatlabControl.